requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.31.0
numpy>=1.24.3
scikit-learn==1.4.1.post1